+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Reverse complement nucleotide colours were wrong, and now nucleotide colours work in comparison mode.
 - Improved matches in comparison mode, previously very low percent identity matches are now split into more accurate higher percent identity matches.
 - Some matches incorrectly reported as low, when actually high. Zooming in showed no match lines in these cases, but now also fixed.
+- Comparison mode bug where matches could be missing at repeats, noticeable in order genomes were added or when trying to compare genome to itself
 
 ## [1.0.0] - "Total Perspective" - 2026-04-05